+++ packages/vue/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Add a `curi-focus` directive to auto-focus an element when new responses are emitted.
+
 ## 1.0.0-beta.18
 
 * Add scoped slot to `<curi-link>` for accessing the link's `navigating` state.